refactor: Refactored file structure adapting to atomic desing standards
