+++ src/ui/_components/index.ts
@@ -1,14 +1,3 @@
-export { Text } from './atoms/Text'
-
-export { Box } from './objects/Box'
-export { Row } from './objects/Row'
-export { CommonGoldInput } from './objects/inputs/CommonGoldInput'
-export { BigGoldInput } from './objects/inputs/BigGoldInput'
-export { GoldSelectBox } from './objects/inputs/GoldSelectBox'
-export { FormWrapper } from './objects/wrappers/section-wrappers/FormWrapper'
-export { ContentWrapper } from './objects/wrappers/section-wrappers/ContentWrapper'
-export { HeaderWrapper } from './objects/wrappers/section-wrappers/HeaderWrapper'
-export { GoldButton } from './objects/buttons/GoldButton'
-export { SectionTitleWrapper } from './objects/wrappers/section-wrappers/SectionTitleWrapper'
-export { DefaultFormElementWrapper } from './objects/wrappers/component-wrappers/DefaultFormElementWrapper'
-export { DefaultButton } from './objects/buttons/DefaultButton'
+export { DefaultFormElementWrapper, Box, Row } from './objects/'
+export { SectionTitle } from './molecules'
+export { Form, Header } from './organisms'
