+++ src/ui/_components/molecules/buttons/FormButton.tsx
@@ -1,0 +1,14 @@
+import { DefaultFormElementWrapper } from '../../'
+import { GoldButton } from '../../atoms'
+
+type buttonProps = {
+  children: JSX.Element | JSX.Element[] | string
+}
+
+export const FormButton = ({ children }: buttonProps) => {
+  return (
+    <DefaultFormElementWrapper>
+      <GoldButton>{children}</GoldButton>
+    </DefaultFormElementWrapper>
+  )
+}
