+++ src/ui/_components/molecules/fillers/SpaceFiller.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled'
+
+export const SpaceFiller = styled.div<{ flex: number }>`
+  flex: ${props => props.flex};
+  height: 100%;
+`
