+++ src/ui/_components/molecules/inputs/BigFormInput.tsx
@@ -1,0 +1,10 @@
+import { DefaultFormElementWrapper } from '../../index'
+import { BigGoldInput } from '../../atoms'
+
+export const BigFormInput = ({ value = '' }: { value: string }) => {
+  return (
+    <DefaultFormElementWrapper>
+      <BigGoldInput value={value} />
+    </DefaultFormElementWrapper>
+  )
+}
