+++ src/ui/_components/molecules/inputs/SmallFormInput.tsx
@@ -1,0 +1,10 @@
+import { CommonGoldInput } from '../../atoms/inputs/CommonGoldInput'
+import { DefaultFormElementWrapper } from '../../'
+
+export const SmallFormInput = ({ value = '' }: { value: string }) => {
+  return (
+    <DefaultFormElementWrapper>
+      <CommonGoldInput value={value} />
+    </DefaultFormElementWrapper>
+  )
+}
