--- src/ui/_components/objects/inputs/inputProps.ts
@@ -1,3 +1,0 @@
-export default interface Props {
-  value?: string
-}
