+++ src/ui/_components/objects/wrappers/DefaultFormElementWrapper.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled'
 
-const DefaultFormElementWrapper = styled.div<{ justifyContent?: string }>`
+const DefaultFormElementWrapper = styled.div<{ justifyContent?: 'flex-start' | 'flex-end' | 'center' }>`
   padding: 20px;
   display: flex;
   justify-content: ${props => props.justifyContent || 'center'};
