+++ src/ui/_components/organisms/Header.tsx
@@ -1,0 +1,26 @@
+import styled from '@emotion/styled'
+import { MainTitle, ReturnsTitle, SpaceFiller } from '../molecules/index'
+
+const HeaderDiv = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  min-height: 20%;
+  width: 100%;
+
+  border-radius: 0 0 27px 27px;
+
+  background-image: url('https://pamplonaswcraft.com/assets/images/bg-hero.jpg');
+  background-size: cover;
+`
+
+export const Header = (): JSX.Element => {
+  return (
+    <HeaderDiv>
+      <SpaceFiller flex={4} />
+      <MainTitle />
+      <ReturnsTitle />
+    </HeaderDiv>
+  )
+}
